refactor: move client types into client dir
