+++ src/client/types/textDocument.ts
