+++ src/environment/environment.ts
@@ -1,9 +1,9 @@
 import { Observable, of } from 'rxjs'
 import { distinctUntilChanged, map } from 'rxjs/operators'
 import { TextDocument } from 'sourcegraph'
+import { TextDocumentItem } from '../client/types/textDocument'
 import { ConfigurationCascade } from '../protocol'
 import { Range, Selection } from '../protocol/plainTypes'
-import { TextDocumentItem } from '../types/textDocument'
 import { isEqual } from '../util'
 import { Context, EMPTY_CONTEXT } from './context/context'
 import { Extension } from './extension'
