+++ src/extension/old/api.ts
@@ -1,9 +1,9 @@
 import { Subscription } from 'rxjs'
 import * as sourcegraph from 'sourcegraph'
+import { TextDocumentIdentifier } from '../../client/types/textDocument'
 import { Context } from '../../environment/context/context'
 import { MessageConnection } from '../../jsonrpc2/connection'
 import { InitializeParams, Settings, TextDocumentDecoration } from '../../protocol'
-import { TextDocumentIdentifier } from '../../types/textDocument'
 
 /**
  * The Sourcegraph extension API, which extensions use to interact with the client.
