+++ src/extension/types/errors.ts
