+++ src/extension/types/location.ts
