+++ src/extension/types/selection.ts
